+++ package.json
@@ -3,7 +3,7 @@
   "productName": "Blowfish",
   "version": "1.0.2",
   "description": "🐡 Simple Electron-based desktop app to retrieve and display your total Ocean Token balances.",
-  "main": "./src/main.js",
+  "main": "./src/main/index.js",
   "scripts": {
     "test": "eslint ./src/**/*.{js,jsx} && stylelint ./src/**/*.css",
     "start": "webpack-dev-server --hot --host 0.0.0.0 --config=./webpack.dev.config.js",
@@ -21,13 +21,8 @@
   "license": "MIT",
   "dependencies": {
     "@coingecko/cryptoformat": "^0.3.1",
-    "@reach/router": "^1.2.1",
     "ethereum-address": "0.0.4",
-    "ms": "^2.1.1",
-    "react": "^16.8.6",
-    "react-blockies": "^1.4.1",
-    "react-dom": "^16.8.6",
-    "react-pose": "^4.0.8"
+    "ms": "^2.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.4.5",
@@ -36,6 +31,7 @@
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.4.5",
+    "@reach/router": "^1.2.1",
     "@svgr/webpack": "^4.2.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
@@ -52,6 +48,10 @@
     "html-webpack-plugin": "^3.2.0",
     "prettier": "^1.17.0",
     "prettier-stylelint": "^0.4.2",
+    "react": "^16.8.6",
+    "react-blockies": "^1.4.1",
+    "react-dom": "^16.8.6",
+    "react-pose": "^4.0.8",
     "style-loader": "^0.23.1",
     "stylelint": "^10.0.1",
     "stylelint-config-standard": "^18.3.0",
@@ -64,6 +64,7 @@
     "appId": "com.kremalicious.blowfish",
     "files": [
       "./build/**/*",
+      "./src/main/**/*",
       "./src/*.js",
       "package.json"
     ],
