+++ src/main/menu.js
@@ -1,6 +1,6 @@
 const { app, Menu } = require('electron')
-const { openUrl } = require('./utils')
-const { homepage } = require('../package.json')
+const { openUrl } = require('../utils')
+const { homepage } = require('../../package.json')
 
 const buildMenu = mainWindow => {
   const template = [
