+++ src/renderer/components/Animations.js
