+++ src/renderer/components/Balance.css
