+++ src/renderer/components/Balance.jsx
