+++ src/renderer/components/Spinner.css
