+++ src/renderer/components/Spinner.jsx
