+++ src/renderer/components/Ticker.css
