+++ src/renderer/components/Ticker.jsx
