+++ src/renderer/components/Total.jsx
