+++ src/renderer/components/Welcome.jsx
