+++ src/renderer/images/icon.icns
