+++ src/renderer/images/icon.ico
