+++ src/renderer/images/icon1024.png
